SDK 6.x compatibility: refactor kuzzle wrapper to enable instansation of a new Kuzzle SDK object while changing environment
