+++ package.json
@@ -84,7 +84,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^2.0.2",
-    "kuzzle-sdk": "^5.0.6",
+    "kuzzle-sdk": "^6.0.0",
     "lolex": "^1.5.2",
     "materialize-css": "0.97.6",
     "mocha": "^3.2.0",
