+++ src/components/Common/Login/Form.vue
@@ -64,7 +64,7 @@
     DO_LOGIN,
     PREPARE_SESSION
   } from '../../../vuex/modules/auth/mutation-types'
-  import kuzzle from '../../../services/kuzzle'
+  import {kuzzle} from '../../../services/kuzzleWrapper'
 
   export default {
     name: 'LoginForm',
