+++ src/components/Data/Documents/Update.vue
@@ -45,8 +45,7 @@
 
   import CollectionDropdown from '../Collections/Dropdown'
   import Headline from '../../Materialize/Headline'
-  import kuzzle from '../../../services/kuzzle'
-  import { getMappingDocument } from '../../../services/kuzzleWrapper'
+  import { kuzzle, getMappingDocument } from '../../../services/kuzzleWrapper'
   import CreateOrUpdate from './Common/CreateOrUpdate'
   import CollectionTabs from '../Collections/Tabs'
   import {SET_TOAST} from '../../../vuex/modules/common/toaster/mutation-types'
