+++ src/components/Security/Roles/Update.vue
@@ -14,8 +14,7 @@
 
 <script>
   import CreateOrUpdate from '../../Data/Documents/Common/CreateOrUpdate'
-  import kuzzle from '../../../services/kuzzle'
-  import { getMappingRoles } from '../../../services/kuzzleWrapper'
+  import { kuzzle, getMappingRoles } from '../../../services/kuzzleWrapper'
   import {SET_TOAST} from '../../../vuex/modules/common/toaster/mutation-types'
 
   export default {
